Separated 'Season' to its own file. Got rid of redundant 'pick_game'
