+++ ReinforcementLearning/NHL/analyses/prediction_Qvalues2Teampoints.py
@@ -12,8 +12,8 @@ from Utils.programming.ut_find_folders import *
 from Utils.programming.ut_clone_directory import  *
 from ReinforcementLearning.NHL.playbyplay.state_space_data import HockeySS
 from ReinforcementLearning.NHL.playerstats.nhl_player_stats import do_ANN_training, do_clustering_multiyear, PlayerStatsFetcher
-from ReinforcementLearning.NHL.playbyplay.playbyplay_data import Season, Game
-
+from ReinforcementLearning.NHL.playbyplay.playbyplay_data import Game
+from ReinforcementLearning.NHL.playbyplay.season import Season
 
 
 def get_players_classes(repoModel, data_for_game, preprocessing, classifier, number_of_games):
